Fixed card issue
For some reason my legal move computer printed "Hare move not implemented" a thousand times.
The problem was, that I added the cards of the hares to their card vector without reseting it.
So if I bought a Card in round 1 then my client would think that I had 20 of them in round 20